prepare for GitHub pages

+++ src/main.py
@@ -1,3 +1,4 @@
+import sys
 import os
 import shutil
 from markdown.conversion import generate_page
@@ -5,10 +6,11 @@ from markdown.conversion import generate_page
 STATIC_PATH = "static"
 CONTENT_PATH = "content"
 TEMPLATE_PATH = "template.html"
-PUBLIC_PATH = "public"
+PUBLIC_PATH = "docs"
 
 def clear_public_dir():
-    shutil.rmtree(PUBLIC_PATH)
+    if os.path.exists(PUBLIC_PATH):
+        shutil.rmtree(PUBLIC_PATH)
 
 def copy_static_files(sub_path=""):
     path = os.path.join(STATIC_PATH, sub_path)
@@ -26,24 +28,30 @@ def copy_static_files(sub_path=""):
                 print(f"copy file {filename}")
                 shutil.copy(file_path, os.path.join(dest_dir, filename))
 
-def generate_pages(sub_path=""):
+def generate_pages(base_path, sub_path=""):
     path = os.path.join(CONTENT_PATH, sub_path)
     if os.path.exists(path):
         filenames = os.listdir(path)
         for filename in filenames:
             file_path = os.path.join(path, filename)
             if os.path.isdir(file_path):
-                generate_pages(os.path.join(sub_path, filename))
+                generate_pages(base_path, os.path.join(sub_path, filename))
             else:
                 dest_filename = filename.removesuffix(".md") + ".html"
                 dest_dir = os.path.join(PUBLIC_PATH, sub_path)
-                generate_page(file_path, TEMPLATE_PATH, os.path.join(dest_dir, dest_filename))
+                generate_page(base_path, file_path, TEMPLATE_PATH, os.path.join(dest_dir, dest_filename))
 
 
 def main():
+    base_path = "/"
+    if (len(sys.argv) > 2):
+        raise Exception("Usage: main.py [basepath]")
+    if (len(sys.argv) == 2):
+        base_path = sys.argv[1]
+
     clear_public_dir()
     copy_static_files()
-    generate_pages()
+    generate_pages(base_path)
 
 if __name__ == "__main__":
     main()
